django/requirements.txt
django==3.1.7
djangorestframework==3.12.2
djangorestframework-gis==0.17
django-filter==2.4.0
django-url-filter==0.3.15
djangorestframework-queryfields==1.0.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.7.0
djangorestframework-mvt==0.2.5
django-allauth==0.44.0
gunicorn==20.0.4
asn1crypto==1.4.0
cffi==1.14.5
cryptography
ecdsa==0.14.1
future==0.18.2
idna==2.10
pycparser==2.20
pycrypto==2.6.1
pyjwt==1.7.1
python-jose==3.2.0
markdown==3.3.4
packaging==20.9
six==1.15.0
psycopg2-binary==2.8.6
python-dotenv==0.15.0 
pytz==2021.1
